# backend/docs/swagger.yaml
definitions:
  dto.AuthTokenDto:
    properties:
      token:
        type: string
    type: object
  dto.CreateGameDto:
    properties:
      description:
        maxLength: 2000
        type: string
      genreId:
        type: string
      releaseDate:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - genreId
    - releaseDate
    - title
    type: object
  dto.CreateGenreDto:
    properties:
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - title
    type: object
  dto.CreateUserDto:
    properties:
      password:
        maxLength: 200
        minLength: 1
        type: string
      userRole:
        $ref: '#/definitions/specifictype.UserRole'
      username:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - password
    - userRole
    - username
    type: object
  dto.GameDto:
    properties:
      description:
        type: string
      genreId:
        type: string
      id:
        type: string
      releaseDate:
        type: string
      title:
        type: string
    type: object
  dto.GenreDto:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
  dto.LoginDto:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.UpdateGameDto:
    properties:
      description:
        maxLength: 2000
        type: string
      genreId:
        type: string
      id:
        type: string
      releaseDate:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - genreId
    - id
    - releaseDate
    - title
    type: object
  dto.UpdateGenreDto:
    properties:
      id:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - id
    - title
    type: object
  dto.UpdateUserDto:
    properties:
      id:
        type: string
      password:
        maxLength: 200
        minLength: 1
        type: string
      userRole:
        $ref: '#/definitions/specifictype.UserRole'
      username:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - id
    - password
    - userRole
    - username
    type: object
  dto.UserDto:
    properties:
      id:
        type: string
      userRole:
        $ref: '#/definitions/specifictype.UserRole'
      username:
        type: string
    type: object
  specifictype.UserRole:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
info:
  contact: {}
  title: Gin Swagger Example
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Принимает логин и пароль и возвращает JWT токен
      parameters:
      - description: Логин и пароль
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthTokenDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация
      tags:
      - auth
  /games:
    get:
      consumes:
      - application/json
      description: Возвращает список всех игр в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GameDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все игры
      tags:
      - games
    post:
      consumes:
      - application/json
      description: Создает новую игру в системе
      parameters:
      - description: Данные для создания игры
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGameDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GameDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать игру
      tags:
      - games
  /games/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет игру из системы по её идентификатору
      parameters:
      - description: ID игры
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить игру
      tags:
      - games
    get:
      consumes:
      - application/json
      description: Получает информацию об игре по её идентификатору
      parameters:
      - description: ID игры
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GameDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить игру
      tags:
      - games
    put:
      consumes:
      - application/json
      description: Обновляет информацию об игре
      parameters:
      - description: ID игры
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления игры
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGameDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GameDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить игру
      tags:
      - games
  /genres:
    get:
      consumes:
      - application/json
      description: Возвращает список всех жанров в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GenreDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все жанры
      tags:
      - genres
    post:
      consumes:
      - application/json
      description: Создает новый жанр в системе
      parameters:
      - description: Данные для создания жанра
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGenreDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GenreDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать жанр
      tags:
      - genres
  /genres/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет жанр из системы по его идентификатору
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить жанр
      tags:
      - genres
    get:
      consumes:
      - application/json
      description: Получает информацию о жанре по его идентификатору
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GenreDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить жанр
      tags:
      - genres
    put:
      consumes:
      - application/json
      description: Обновляет информацию о жанре
      parameters:
      - description: ID жанра
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления жанра
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGenreDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GenreDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить жанр
      tags:
      - genres
  /health:
    get:
      consumes:
      - application/json
      description: Проверяет, что сервис работает
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверка здоровья
      tags:
      - health
  /users:
    get:
      consumes:
      - application/json
      description: Возвращает список всех пользователей в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить всех пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные для создания пользователя
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет пользователя из системы по его идентификатору
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить пользователя
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Получает информацию о пользователе по его идентификатору
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет информацию о пользователе
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления пользователя
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить пользователя
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите значение заголовка целиком: "Bearer <JWT>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"